% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/holdout.core.R
\name{createRandomSet}
\alias{createRandomSet}
\title{Create a set of historical data and test data.}
\usage{
createRandomSet(num.features, num.samples, nbiased = 20,
  bias = 6/sqrt(num.features))
}
\arguments{
\item{num.features}{- number of features in each dataset}

\item{num.samples}{- number of samples in each dataset}

\item{nbiased}{- number of features with a true correlation
between feature value and signal value}

\item{bias}{- a factor that determines the strength of
correlations for the nbiased features}
}
\description{
This function will output a list with four objects.
Two objects will be data matrices representing historical
and test data. Two objects will be signals or sample classes
that describe the samples in both datasets.
}

