% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/holdout.core.R
\name{evaluatePredictionAccuracy}
\alias{evaluatePredictionAccuracy}
\title{Compute the accuracy comparing a set of truth values and predictions}
\usage{
evaluatePredictionAccuracy(truth, prediction)
}
\arguments{
\item{truth}{- a vector holding true labels}

\item{prediction}{- a vector holding labels output by some
prediction method}
}
\description{
Compare a vector of truth classes and prediction classes.
The function is actually generic. It works for +1/-1/0 classes,
by any other labels work as well.
}

