% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/holdout.core.R
\name{getBinaryFeatureWeights}
\alias{getBinaryFeatureWeights}
\title{Attribute weights to each feature according to correlations values}
\usage{
getBinaryFeatureWeights(cors, k)
}
\arguments{
\item{cors}{- vector or matrix with correlations
If it is a vector, it is intepreted as containign correlations for
each feature. If a matrix, it is intepreted as having features as rows,
and various estimates in columns. The function will use the mean
consensus correlation among the columns.}

\item{k}{- number of features to call}
}
\description{
Function outputs a vector with values in {-1, 0, 1} determining
binary weights for each element in the cors vector.
Technically, the weights are not binary, bu
}

