% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/holdout.core.R
\name{getFeatureCorrelations}
\alias{getFeatureCorrelations}
\title{Compute correlations between a data matrix and a signal vector}
\usage{
getFeatureCorrelations(dat, signal, min.cor = 1/sqrt(length(signal)),
  dat.bg = NULL, signal.bg = NULL, tolerance.factor = 1,
  threshold.factor = 4)
}
\arguments{
\item{dat}{- a data matrix with S samples in columns and F features
rows}

\item{signal}{- a numeric vector. The function will compute
correlations between rows in the data matrix and this signal vector.}

\item{min.cor}{- minimal expected correlation (function
will output zero if the actuall correlation is below
threshold)}

\item{dat.bg}{- background dataset}

\item{signal.bg}{- background signal}

\item{tolerance.factor}{- one of the penalties used in the
reusable holdout proposal}

\item{threshold.factor}{- one of the penalties used in the
reusable holdout proposal.}
}
\description{
Evaluates correlations between features in a data matrix and a
signal vector. When only one data and signal object is provided,
the output is a vector of straightforward correlations. When a
background dataset and background signal vector are also provided,
the function treats the primary data as a reusable holdout: it
will output correlations from either the holdout data or the
background set.
}

