% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoldoutMetrics.R
\docType{class}
\name{HoldoutMetrics}
\alias{HoldoutMetrics}
\title{HoldoutMetrics}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HoldoutMetrics
}
\description{
Create HoldoutMetrics object
}
\section{Methods}{

\describe{
  \item{\code{$new(metric, train_target, holdout_target, threshold, sigma, budget)}}{Creates a new \code{Thresholdout} object.}
  \item{\code{$query(train_pred, holdout_pred)}}{Sends a query to the holdout dataset.}
}
}

\section{Actives}{

\describe{None. See Thresholdout.}
}

\examples{
library(rThresholdout)
library(ModelMetrics)

# split data
data(testDF)
train <- testDF[1:50,]
holdout <- testDF[51:100,]

# identify targets
train_target <- train$y
holdout_target <- holdout$y

# build a model
model <- glm(y ~ ., data = train, family="binomial")

# generate predictions
train_preds <- predict(model, train)
holdout_preds <- predict(model, holdout)

# ModelMetrics provides a consistent Rcpp implementation for several measures of fit
train_metrics <- auc(actual = train_target, predicted = train_preds)
holdout_metrics <- auc(actual = holdout_target, predicted = holdout_preds)

# Thresholdout returns metrics derived from train and holdout data sets.
Thresholdout_Obj <- Thresholdout$new(Threshold = 0.04, Sigma = 0.01, Budget = 1000)
Thresholdout_Obj$query(train_val = train_metrics, holdout_val = holdout_metrics)


# HoldoutMetrics simplifies this process
HoldoutMetrics_Obj <- HoldoutMetrics$new(metric = auc,
                                         train_target = train_target,
                                         holdout_target = holdout_target,
                                         threshold = 0.04,
                                         sigma = 0.01,
                                         budget = 1000)

HoldoutMetrics_Obj$query(train_preds,
                         holdout_preds)

}
\keyword{datasets}
