% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/holdout.core.R
\name{predictFromWeights}
\alias{predictFromWeights}
\title{Simple prediction of a +1,-1,0 class based on a data matrix
and feature weights}
\usage{
predictFromWeights(dat, weights)
}
\arguments{
\item{dat}{- a data matrix with samples in columns and features
in rows}

\item{weights}{- a numeric vector with weights for each feature.}
}
\description{
Simple prediction of a +1,-1,0 class based on a data matrix
and feature weights
}

