% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rThresholdout.R
\docType{package}
\name{rThresholdout}
\alias{rThresholdout}
\alias{rThresholdout-package}
\title{rThresholdout: Thresholdout Algorithm}
\description{
A Pure R implementation of Thresholdout Algorithm for Adaptive Data Analysis
and Holdout Reuse. This algorithm was proposed in Dwork et al. (2015)
\url{https://arxiv.org/abs/1506.02629}.
}

