% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholdout.R
\docType{class}
\name{Thresholdout}
\alias{Thresholdout}
\title{Thresholdout}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Thresholdout
}
\description{
Create Thresholdout object for adaptive data analysis and holdout reuse.
}
\section{Methods}{

\describe{
  \item{\code{$new(threshold, sigma, budget)}}{Creates a new \code{Thresholdout} object.}
  \item{\code{$query(train_val, holdout_val)}}{Sends a query to the holdout dataset.}
}
}

\section{Actives}{

\describe{
  \item{\code{$threshold}}{gets the threshold value.}
  \item{\code{$sigma}}{gets the sigma value.}
  \item{\code{$budget}}{gets the remaining budget value.}
  \item{\code{$record}}{gets the record of previous queries.}
}
}

\examples{
library(rThresholdout)
Thresholdout_Obj <- Thresholdout$new(threshold = 0.04, sigma = 0.01, budget = 1000)
Thresholdout_Obj$query(train_val = c(0.06, 0.07), holdout_val = c(0.05, 0.08))

}
\keyword{datasets}
