% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholdout.R
\docType{class}
\name{Thresholdout}
\alias{Thresholdout}
\title{Thresholdout}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Thresholdout
}
\description{
Create Thresholdout object for adaptive data analysis and holdout reuse.
}
\section{Methods}{

\describe{
  \item{\code{$new(Threshold, Sigma, Budget)}}{Creates a new \code{Thresholdout} object.}
  \item{\code{$query(train_val, holdout_val)}}{Sends a query to the holdout dataset.}
}
}

\section{Actives}{

\describe{
  \item{\code{$Threshold}}{gets the Threshold value.}
  \item{\code{$Sigma}}{gets the Sigma value.}
  \item{\code{$Budget}}{gets the remaining Budget value.}
  \item{\code{$Record}}{gets the Record of previous queries.}
}
}
\examples{
library(rThresholdout)
Thresholdout_Obj <- Thresholdout$new(Threshold = 0.04, Sigma = 0.01, Budget = 1000)
Thresholdout_Obj$query(train_val = c(0.06, 0.07), holdout_val = c(0.05, 0.08))

}
\keyword{datasets}

